Reduced ElevenLabs latency by switching to streaming
Issues: Works in Editor, crashes in build, and I can't for the life of
me figure out why. The error thrown is a memory access violation.

My best guess is that it has something to do with creating AudioClips, a
managed object, on a thread that isn't the main thread.

If this is fixed, it could potentially make interrupts viable as we can
stream the spoken transcript, figuring out what the user has heard so
far.